segmentの実装

+++ src/routes/produce.rs
@@ -7,7 +7,7 @@ use crate::record::{Log, Record};
 pub async fn produce(
     extract::Json(payload): extract::Json<Record>,
     Extension(log): Extension<Arc<Log>>,
-) -> Result<Json<usize>, (StatusCode, String)> {
+) -> Result<Json<u64>, (StatusCode, String)> {
     let offset = log
         .append(payload)
         .map_err(|e| (StatusCode::BAD_REQUEST, format!("{}", e)))?;
